+++ packages/minimal-shared/src/hooks/use-popover-hover/use-popover-hover.ts
@@ -35,18 +35,18 @@ import { useRef, useState, useCallback } from 'react';
 
 type UsePopoverHoverReturn<T> = {
   open: boolean;
-  onOpen: () => void;
   anchorEl: T | null;
+  onOpen: () => void;
   onClose: () => void;
-  elementRef: RefObject<T>;
+  elementRef: RefObject<T | null>;
   setOpen: Dispatch<SetStateAction<boolean>>;
 };
 
 export function usePopoverHover<T extends HTMLElement>(
   inputRef?: RefObject<T>
 ): UsePopoverHoverReturn<T> {
-  const initialRef = useRef<T | null>(null);
-  const elementRef = inputRef || initialRef;
+  const localRef = useRef<T>(null);
+  const elementRef = inputRef || localRef;
 
   const [open, setOpen] = useState<boolean>(false);