+++ packages/minimal-shared/CHANGELOG.md
@@ -1,5 +1,17 @@
 # minimal-shared
 
+## 1.0.6
+
+_Feb 23, 2025_
+
+- Update `hooks/use-scroll-offset-top`.
+- Update `hooks/use-client-rect`.
+- Update `hooks/use-popover-hover`.
+- Update `utils/refs`.
+- Clean devDependencies
+
+---
+
 ## 1.0.5
 
 _Jan 1, 2025_